requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
python-dotenv
httpx
playwright
google-genai
openai
pytest-playwright
networkx